# docs/swagger.yaml
openapi: 3.0.3
info:
  title: TOKO-ENCOUNT
  version: 0.0.1
  description:  TOKO-ENCOUNT API
servers:
  - url: http://localhost:3050
paths:
...
  '/api/chat':
    get:
      tags:
        - chat
      summary: 参加中のチャット一覧を表示
      operationId: getChat
      parameters:
        - name: limit
          in: query
          description: 取得するチャットの最大数
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: 何件目のチャットから取得するか
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: チャット情報
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatList"
    post:
      tags:
        - chat
      summary: 新規にプライベートルームを作成
      operationId: createChat
      parameters:
        - name: did
          in: query
          required: true
          description: トーク相手のユーザーID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: メッセージ情報
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatData"
  /api/chat/{rid}:
    get:
      tags:
        - chat
      summary: ridのルーム内のメッセージを取得
      operationId: getChatMessages
      parameters:
        - name: rid
          in: path
          required: true
          description: ルームID(全体チャットは0)
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: 取得するメッセージの最大数
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: 何件目のメッセージから取得するか
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: メッセージ情報
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
    post:
      tags:
        - chat
      summary: ridにメッセージを送信
      operationId: postChat
      parameters:
        - name: rid
          in: path
          required: true
          description: ルームID(全体チャットは0)
          schema:
            type: string
            format: uuid
        - name: did
          in: query
          required: true
          description: 送信相手のユーザーID(全体チャットは0)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                post:
                  type: string
      responses:
        '200':
          description: メッセージ送信に成功しました
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
components:
  schemas:
...
    Message:
      type: object
      properties:
        chatId:
          type: string
          format: uuid
        post:
          type: string
          example: hello
        postUserId:
          type: string
          format: uuid
        userName:
          type: string
          example: testUserName
        createdAt:
          type: string
          format: date-time
      required:
        - chatId
        - post
        - postUserId
        - userName
        - createdAt
    MessageList:
      type: object
      properties:
        hasNext:
          type: boolean
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required:
        - hasNext
        - messages
    ChatData:
      type: object
      properties:
        roomId:
          type: string
        name:
          type: string
        latestMessage:
          $ref: "#/components/schemas/Message"
        newMessageCount:
          type: integer
      required:
        - roomId
        - name
        - latestMessage
        - newMessageCount
    ChatList:
      type: object
      properties:
        hasNext:
          type: boolean
        chats:
          type: array
          items:
            $ref: "#/components/schemas/ChatData"
      required:
        - hasNext
        - chats